chore: upgrade dep, tweak test runner, migrate to gh action
