+++ .github/ISSUE_TEMPLATE.md
@@ -1,0 +1,76 @@
+<!--
+BUGS: Please use this template.
+
+SUPPORT REQUESTS/QUESTIONS: If you have a support request or question please
+submit them to StackOverflow using the tags aurelia and aurelia-framework
+http://stackoverflow.com/questions/tagged/aurelia
+or the Aurelia Gitter https://gitter.im/aurelia/discuss
+
+Blue Spire offers paid support agreements. Further information regarding paid support
+may be obtained by emailing support@bluespire.com
+
+Future support requests submitted here will be closed.
+
+The HTML comments below are for your reference, and are not displayed
+when your issue is submitted, feel free to leave them.
+
+Choose one of the two headings, delete the other.
+-->
+**I'm submitting a bug report**
+**I'm submitting a feature request**
+
+* **Library Version:**
+major.minor.patch-pre
+
+
+**Please tell us about your environment:**
+* **Operating System:**
+OSX 10.x|Linux (distro)|Windows [7|8|8.1|10]
+
+* **Node Version:**
+6.2.0
+<!--
+Minimum supported Node version is latest Node 4.x LTS
+run `node -v`
+-->
+
+* **NPM Version:**
+3.8.9
+<!--
+Minimum supported NPM version is 3.x
+run `npm -v`
+-->
+
+* **JSPM OR Webpack AND Version**
+JSPM 0.16.32 | webpack 2.1.0-beta.17
+<!--
+If using JSPM
+run `jspm -v`
+If using Webpack
+run `webpack --help | grep webpack`
+-->
+
+* **Browser:**
+all | Chrome XX | Firefox XX | Edge XX | IE XX | Safari XX | Mobile Chrome XX | Android X.X Web Browser | iOS XX Safari | iOS XX UIWebView | iOS XX WKWebView
+
+* **Language:**
+all | TypeScript X.X | ESNext
+
+
+**Current behavior:**
+
+
+**Expected/desired behavior:**
+<!--
+If the current behavior is a bug, please provide the steps to reproduce and, if possible, a minimal demo of the
+problem along with a runnable gist, if possible.
+To create a runnable gist, go to https://gist.run/?id=7542e061bc940cde506b&sha=6821c521a6c7bae6f59a36fb8628ecd1032b2d10
+Then click "Fork to Public Gist".
+Create your gist, then finally click "Update Gist." Include a link to the gist.run below.
+-->
+
+
+* **What is the expected behavior?**
+
+
+* **What is the motivation / use case for changing the behavior?**
