+++ .github/workflows/main.yml
@@ -1,0 +1,16 @@
+name: main
+on: [push]
+
+jobs:
+
+  ci:
+    timeout-minutes: 10
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: actions/setup-node@v1
+        with:
+          node-version: 14
+      - run: npm ci
+      - run: npm run test
+      - run: npm run cut-release
