+++ bower.json
@@ -1,6 +1,6 @@
 {
   "name": "aurelia-validation",
-  "version": "2.0.0-rc1",
+  "version": "2.0.0",
   "description": "Validation for Aurelia applications",
   "keywords": [
     "aurelia",
