+++ doc/CHANGELOG.md
@@ -1,3 +1,9 @@
+# Changelog
+
+All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+
+## [2.0.0](https://github.com/aurelia/validation/compare/2.0.0-rc2...2.0.0) (2023-02-02)
+
 # [2.0.0-rc2](https://github.com/aurelia/validation/compare/2.0.0-rc1...2.0.0-rc2) (2020-06-12)
 
 ## Bug Fixes
