+++ package.json
@@ -31,46 +31,20 @@
     "LICENSE"
   ],
   "scripts": {
-    "lint": "cross-env tslint --project tsconfig.json",
-    "pretest": "cross-env npm run lint",
-    "test": "cross-env rimraf dist/test dist/src && tsc && karma start --single-run",
-    "develop": "concurrently \"./node_modules/.bin/tsc --watch\" \"./node_modules/.bin/karma start\"",
-    "prebuild": "cross-env rimraf dist",
+    "lint": "eslint .",
+    "test": "karma start --single-run",
+    "test:watch": "npm run test -- --no-single-run",
+    "test:debugger": "npm run test -- --browsers ChromeDebugging --no-single-run",
     "build": "rollup -c --environment NODE_ENV:production",
     "postbuild": "npm run build:dts",
     "build:dts": "dts-bundle-generator -o dist/aurelia-validation.d.ts src/aurelia-validation.ts",
     "predoc": "cross-env rimraf doc/api.json && rimraf dist/doc-temp && tsc --project tsconfig.build.json --outFile dist/doc-temp/aurelia-validation.js && node doc/shape-defs && copyfiles tsconfig.json dist/doc-temp",
-    "doc": "cross-env typedoc --json doc/api.json --excludeExternals --includeDeclarations --mode modules --target ES6 --name aurelia-validation-docs dist/doc-temp/",
-    "postdoc": "cross-env node doc/shape-doc && rimraf dist/doc-temp",
-    "precut-release": "cross-env npm run test",
-    "changelog": "conventional-changelog -p angular -i doc/CHANGELOG.md -s",
-    "cut-release": "standard-version -t \"\" -i doc/CHANGELOG.md && npm run build && npm run doc"
-  },
-  "jspm": {
-    "registry": "npm",
-    "jspmPackage": true,
-    "main": "aurelia-validation",
-    "format": "amd",
-    "directories": {
-      "dist": "dist/amd"
-    },
-    "peerDependencies": {
-      "aurelia-binding": "^2.5.2",
-      "aurelia-dependency-injection": "^1.5.0",
-      "aurelia-logging": "^1.5.0",
-      "aurelia-pal": "^1.8.0",
-      "aurelia-task-queue": "^1.3.1",
-      "aurelia-templating": "^1.9.0"
-    },
-    "dependencies": {
-      "aurelia-binding": "^2.5.2",
-      "aurelia-dependency-injection": "^1.4.1",
-      "aurelia-logging": "^1.5.0",
-      "aurelia-pal": "^1.8.0",
-      "aurelia-task-queue": "^1.3.1",
-      "aurelia-templating": "^1.9.0"
-    },
-    "devDependencies": {}
+    "doc": "typedoc --json doc/api.json --excludeExternals --name aurelia-validation-docs src/aurelia-validation.ts",
+    "postdoc": "node doc/cleanup.js",
+    "precut-release": "npm run test && npm run lint && npm run build",
+    "changelog": "standard-version -t \"\" -i doc/CHANGELOG.md",
+    "cut-release": "npm run changelog",
+    "postcut-release": "npm run doc"
   },
   "dependencies": {
     "aurelia-binding": "^2.5.2",
@@ -78,37 +52,44 @@
     "aurelia-logging": "^1.5.0",
     "aurelia-pal": "^1.8.0",
     "aurelia-task-queue": "^1.3.1",
-    "aurelia-templating": "^1.9.0"
+    "aurelia-templating": "^1.11.1"
   },
   "devDependencies": {
-    "@types/jsdom": "^12.2.0",
-    "@types/node": "^8.5.1",
-    "aurelia-bootstrapper": "^2.3.0",
-    "aurelia-pal-browser": "^1.8.0",
-    "aurelia-pal-nodejs": "^1.1.1",
-    "aurelia-polyfills": "^1.3.0",
-    "aurelia-testing": "^1.0.0",
-    "concurrently": "^4.0.1",
-    "conventional-changelog-cli": "^2.0.5",
-    "copyfiles": "^2.1.0",
-    "cross-env": "^5.2.0",
-    "dts-bundle-generator": "^1.6.1",
-    "jasmine-core": "^3.2.1",
-    "jspm": "^0.16.53",
-    "karma": "^3.0.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-ie-launcher": "^1.0.0",
-    "karma-jasmine": "^1.1.2",
-    "karma-requirejs": "^1.1.0",
-    "requirejs": "^2.3.6",
-    "requirejs-text": "^2.0.15",
-    "rimraf": "^2.6.2",
-    "rollup": "^0.66.2",
-    "rollup-plugin-typescript2": "^0.25.3",
-    "standard-version": "^5.0.2",
-    "tslib": "^1.10.0",
-    "tslint": "^5.11.0",
-    "typedoc": "^0.17.1",
-    "typescript": "^3.8.3"
+    "@types/estree": "^1.0.0",
+    "@types/jasmine": "^4.3.1",
+    "@types/jsdom": "^20.0.1",
+    "@types/node": "^18.11.18",
+    "@types/webpack-env": "^1.18.0",
+    "@typescript-eslint/eslint-plugin": "^5.50.0",
+    "@typescript-eslint/parser": "^5.50.0",
+    "aurelia-bootstrapper": "^2.4.0",
+    "aurelia-pal-browser": "^1.8.1",
+    "aurelia-pal-nodejs": "3.0.0-rc.1",
+    "aurelia-polyfills": "^1.3.4",
+    "aurelia-testing": "^1.1.0",
+    "aurelia-webpack-plugin": "^5.0.5",
+    "conventional-changelog-cli": "^2.2.2",
+    "copyfiles": "^2.4.1",
+    "cross-env": "^7.0.3",
+    "dts-bundle-generator": "^7.2.0",
+    "eslint": "^8.33.0",
+    "html-loader": "^4.2.0",
+    "jasmine": "^4.5.0",
+    "jasmine-core": "^4.5.0",
+    "karma": "^6.4.1",
+    "karma-chrome-launcher": "^3.1.1",
+    "karma-jasmine": "^5.1.0",
+    "karma-sourcemap-loader": "^0.3.8",
+    "karma-typescript": "^5.5.3",
+    "karma-webpack": "^5.0.0",
+    "rimraf": "^4.1.2",
+    "rollup": "^2.79.1",
+    "rollup-plugin-typescript2": "^0.34.1",
+    "standard-version": "^9.5.0",
+    "ts-loader": "^9.4.2",
+    "tslib": "^2.5.0",
+    "typedoc": "^0.23.24",
+    "typescript": "^4.9.5",
+    "webpack": "^5.75.0"
   }
 }
