+++ test/resources/index.ts
@@ -1,12 +1,19 @@
 import { FrameworkConfiguration } from 'aurelia-framework';
+import { CustomInput } from './custom-input';
+import { NullableObjectForm } from './nullable-object-form';
+import { NumberValueCustomAttribute, NumberInputCustomElement } from './number-value';
+import { RegistrationForm } from './registration-form';
+import { TriggerForm } from './trigger-form';
+import { ValidationErrorsFormOne } from './validation-errors-form-one';
 
 export function configure(config: FrameworkConfiguration) {
   config.globalResources([
-    './number-value',
-    './registration-form',
-    './trigger-form',
-    './validation-errors-form-one',
-    './nullable-object-form',
-    './custom-input'
+    NumberValueCustomAttribute,
+    NumberInputCustomElement,
+    RegistrationForm,
+    TriggerForm,
+    ValidationErrorsFormOne,
+    NullableObjectForm,
+    CustomInput,
   ]);
 }
